+++ tuikit/src/widget/mod.rs
@@ -1,10 +1,11 @@
+use crossterm::event::Event;
+
 pub use self::align::*;
 // Various pre-defined widget that implements Draw
 pub use self::split::*;
 pub use self::stack::*;
 pub use self::win::*;
 use crate::draw::Draw;
-use crate::event::Event;
 use std::cmp::min;
 mod align;
 mod split;
@@ -15,14 +16,14 @@ mod win;
 /// Whether fixed size or percentage
 #[derive(Debug, Copy, Clone, Default)]
 pub enum Size {
-    Fixed(usize),
-    Percent(usize),
+    Fixed(u16),
+    Percent(u16),
     #[default]
     Default,
 }
 
 impl Size {
-    pub fn calc_fixed_size(&self, total_size: usize, default_size: usize) -> usize {
+    pub fn calc_fixed_size(&self, total_size: u16, default_size: u16) -> u16 {
         match *self {
             Size::Fixed(fixed) => min(total_size, fixed),
             Size::Percent(percent) => min(total_size, total_size * percent / 100),
@@ -31,23 +32,23 @@ impl Size {
     }
 }
 
-impl From<usize> for Size {
-    fn from(size: usize) -> Self {
+impl From<u16> for Size {
+    fn from(size: u16) -> Self {
         Size::Fixed(size)
     }
 }
 
 #[derive(Copy, Clone, Debug)]
 pub struct Rectangle {
-    pub top: usize,
-    pub left: usize,
-    pub width: usize,
-    pub height: usize,
+    pub top: u16,
+    pub left: u16,
+    pub width: u16,
+    pub height: u16,
 }
 
 impl Rectangle {
     /// check if the given point(row, col) lies in the rectangle
-    pub fn contains(&self, row: usize, col: usize) -> bool {
+    pub fn contains(&self, row: u16, col: u16) -> bool {
         if row < self.top || row >= self.top + self.height {
             false
         } else {
@@ -57,7 +58,7 @@ impl Rectangle {
 
     /// assume the point (row, col) lies in the rectangle, adjust the origin to the rectangle's
     /// origin (top, left)
-    pub fn relative_to_origin(&self, row: usize, col: usize) -> (usize, usize) {
+    pub fn relative_to_origin(&self, row: u16, col: u16) -> (u16, u16) {
         (row - self.top, col - self.left)
     }
 
@@ -75,7 +76,7 @@ impl Rectangle {
 pub trait Widget<Message = ()>: Draw {
     /// the (width, height) of the content
     /// it will be the hint for layouts to calculate the final size
-    fn size_hint(&self) -> (Option<usize>, Option<usize>) {
+    fn size_hint(&self) -> (Option<u16>, Option<u16>) {
         (None, None)
     }
 
@@ -95,7 +96,7 @@ pub trait Widget<Message = ()>: Draw {
 }
 
 impl<Message, T: Widget<Message>> Widget<Message> for &T {
-    fn size_hint(&self) -> (Option<usize>, Option<usize>) {
+    fn size_hint(&self) -> (Option<u16>, Option<u16>) {
         (*self).size_hint()
     }
 
@@ -109,7 +110,7 @@ impl<Message, T: Widget<Message>> Widget<Message> for &T {
 }
 
 impl<Message, T: Widget<Message>> Widget<Message> for &mut T {
-    fn size_hint(&self) -> (Option<usize>, Option<usize>) {
+    fn size_hint(&self) -> (Option<u16>, Option<u16>) {
         (**self).size_hint()
     }
 
@@ -123,7 +124,7 @@ impl<Message, T: Widget<Message>> Widget<Message> for &mut T {
 }
 
 impl<Message, T: Widget<Message> + ?Sized> Widget<Message> for Box<T> {
-    fn size_hint(&self) -> (Option<usize>, Option<usize>) {
+    fn size_hint(&self) -> (Option<u16>, Option<u16>) {
         self.as_ref().size_hint()
     }
